+++ CHANGELOG.md
@@ -9,7 +9,8 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 - ...
 ### Changed
-- ...
+- Add `logger.Logger` to `command.ErrorRecorder` to report errors when appending Command failures to the Event Store.
+- `command.ErrorRecorder` must be passed by reference to implement `command.Handler` interface now (size of the struct increased).
 
 ### Deprecated
 - ...